+++ backend/twitter-bot/db/firestore.js
@@ -1,0 +1,6 @@
+const admin = require("firebase-admin");
+admin.initializeApp();
+
+const db = admin.firestore();
+
+exports.db = db;
